fix(ci): gate MLX tests correctly — unbreak SPM + Metal CI jobs
Two CI jobs went red when the DeepSeek port + embeddings landed:

SPM Build & Test (process abort, exit 1):
- testSanitizeRoundTrip claimed to be Metal-free, but instantiating
  DeepseekV32Model (and the arr() helper) creates real MLXArrays —
  first stream init fatally aborts without default.metallib. Now
  gated with requireMLXRuntimeOrSkip.
- factoryInstantiatesDeepseekV32FromConfig (swift-testing) has the
  same instantiation; gated with .enabled(if: mlxMetallibIsAvailable).
  The two registry-only tests stay ungated.

Metal Tests (numeric divergence):
- GitHub macos-26 runners expose paravirtualized Metal; strict 1e-4
  parity vs Python fixtures diverges there (attention prefill/sparse,
  decoder layer, full model) while passing on real Apple Silicon.
  New requireTrustworthyMetalOrSkip() + TEST_RUNNER_MACMLX_UNTRUSTED_METAL=1
  in the CI job skips the six strict-parity suites on the VM; parity
  stays enforced locally and on any future self-hosted runner.
  Behavioral MLX tests (LoRA, cache, sanitize, indexer top-k, speed
  sampler) still run in the Metal job.

Also hoists metallibIsAvailable out of EmbeddingPoolingTests into the
shared MLXTestSupport (mlxMetallibIsAvailable).

Verified locally under all three CI-equivalent conditions:
bare swift test (0 MLX errors, 76 XCTest/19 skips + 203 swift-testing
passed), TEST_RUNNER_MACMLX_UNTRUSTED_METAL=1 xcodebuild (SUCCEEDED,
parity skipped, behavioral tests ran), plain xcodebuild (SUCCEEDED,
all 11 parity tests executed and passed).
